fix: resolve issues #254 #255 #264 #265
- #254: replace CompressionStream with pako for gzip compression
- #255: add Header.tsx with next/image + blur placeholder, update Layout
- #264: UserFunnelChart already implemented (5 stages, conversion, drill-down)
- #265: extract ExperimentCard component, refactor experiments page
