+++ analytics/package.json
@@ -18,6 +18,7 @@
     "lucide-react": "^1.9.0",
     "mapbox-gl": "^3.22.0",
     "next": "15.4.6",
+    "pako": "^2.1.0",
     "papaparse": "^5.5.3",
     "react": "19.1.0",
     "react-chartjs-2": "^5.3.0",
@@ -30,6 +31,7 @@
     "@next/bundle-analyzer": "^15.4.6",
     "@types/mapbox-gl": "^3.5.0",
     "@types/node": "^20",
+    "@types/pako": "^2.0.3",
     "@types/papaparse": "^5.3.16",
     "@types/react": "^19",
     "@types/react-dom": "^19",
